allow correct size allocation for data views (#257)
Implement a mechanism which allows to predict the number of bytes needed to serialize any of the data-structures used by the adblocker, ahead of time (before serialization). This allows to lift the limitation of size completely (beforehand, we had to allocate a safe amount of memory to be sure there would be enough space). As a benefit, only the required amount of memory is used during initialization and updates, and there is no longer an arbitrary and hard-coded upper limit.

A few consequences:

* most classes now expose a `getSerializedSize(...)` method which returns the number of bytes needed to serialize this particular instance.
* the internal structure of `ReverseIndex` has been optimized so that the memory layout is now closer to access patterns (there is a forward progression in the way values are read from memory at runtime)
* documentation has been improved to better explain the way internal data-structures work and why they are implemented in this specific way.
* the bug which prevented us from using `subarray(...)` to deserialize bytes is now resolved and the un-necessary copies (using `slice(...)`) are gone!

Fixes #257 
Fixes #91