create new @cliqz/adblocker-content package with common utils (#264)
We currently rely on rollup to create a small bundle for content
related code imported from @cliqz/adblocker. Multiple times in
the past the bundler was not aggressive enough and code from
background was pulled in content bundles. To make sure we do not
have this issue again, all these content-scripts helpers are moved
into their own package.