Various small fixes. (#746)
* Make sure that all unsupported procedural selectors from cosmetic
filters are dropped to ensure that we only inject valid CSS selectors.
* Fix matching of `domain=` option for domain filters in cases where
specified domain is a subdomain instead of full hostname or full domain.
* Fix partyness detection for requests without a valid domain (but
having a valid hostname). This fixes matching against localhost request
(for instance).
* Fix engine updates stress test which allows to replay all day-to-day
diffs since the beginning of times... (currently about a year) and make
sure that all updates work and resulting engine is byte-identical with
diff-update or full initialization.
* Fix script to analyze size of serialized engines for all presets as
well as all kinds of compression (i.e. none, gzip and brotli). This
allows to keep track of final size after small-strings compression was
applied.