Proof-of-concept how to block YouTube ads.
Background: the current mechanism to inject scriptlets by
asynchronously responding to messages from the content script
is not reliable. Because of the asynchronous communication,
it is possible that the website already executed and the
patching is too late.

A more reliable approach is to use the "tabs.executeScript" API
and to use the webNavigation.onCommitted hook to trigger it.