New option "enablePushInjectionsOnNavigationEvents":
The original method of having the content script pull the scriptlets
introduces extra latency, which becomes a problem on Chrome-based browsers.
A prominent example is YouTube, where the scriptlets sometimes get
executed when the site's JavaScript has already run.

A more reliable mechanism is to use "webNavigation.onCommitted" API
as a hook and "push" the injections through the "tabs.executeScript" API.

The new option is opt-in, but according to the tests it is recommended
for clients to enable it at least for Chromium-based browsers. On Firefox,
there is currently no proof that it is required to use it.