Scriptlet argument escaping
Scriptlets comming from CDN maybe have escaped arguments for sake of backward
compatibility with older adblocker engines - this makes them hard to
read and debug, so they should be unescaped before injection.

Scriptlets arguments that have escape sequences require the escape
charactes to be escaped, otherwise when injected they would be evaluated
and the scriptlet meaning wont be the same.