feat: dynamic identifiers
This adds a capability of runtime evaluated preprocessor identifiers while keeping backward compatibility and not breaking existing syntaxes. With this feature, we're able to have a fine grained control for filter distribution similar to browser feature flags.

However, we should limit the use of these dynamic identifiers not to underestimate other components like config and product-level AB feature control. For an instance, deciding a criteria for AB testing at the library level doesn't make sense.