feat: Optional support for native :has pseudo classes (#4455)
As :has pseudo classes are now supported by the majority of modern browsers it is fine to run filters with :has selectors as regular CSS selectors. But to be on the safe side, we do inject them in separate CSS blocks.

It is important to keep in mind, by default and in Ghostery extension extended selectors are not enabled. This change, loads extended filters which are :has only, but does not inject them as extended selector but rather like regular css selectors.

Co-authored-by: Philipp Claßen <philipp.classen@posteo.de>
Co-authored-by: HoJeong Go  <seia@outlook.kr>