fix(lint): @typescript-eslint/no-floating-promises
`void` is used to ignore floating promises warning for safe cases. Mostly, they should be an entrypoint function (often called as "main") and errors will be thrown as "unhandledRejection" which is fine for them. The another case would be to skip awaiting on specific promise where awaiting is not wanted. We probably want a proper error handling for those, but discussion should be made to do that. For those, using all of "await", "then" (to chain the output if required), and "catch" keyword is pretty handy solution since we can only see the return value in the code flow: `const x = await y().catch(e => false); if (x === false) throw 'promise failed!'`.

It worth noting that some files (e.g. adblocker-puppeteer and adblocker-playwright) including event listener, they were put on the top. It has no effect at the end but it allows us to use `await` on desired point. Even if putting `await` for those functions are ineffective, it makes the promise to be appear and recognized explicitly.