fix: @typescript-eslint/no-redundant-type-constituents
All changes are regarded safe.

For `unknown`, the problem happens as we abstract the type with generic. It's no longer `unknown` and I added a coverage for undefined values.

For `any`, it just includes all other types, and it's not generic. Merging `undefined` to `any` doesn't matter in this case.