fix(lint): @typescript-eslint/no-unsafe-assignment
The removal of this eslint rule ignorance is actually critical since it means that we allowed using `any` in TypeScript so far. We need to consider using validation libraries like TypeBox for extra safety and performance (the validation speed is forte with compilation time optimization and the overhead is low as seamless level). For the existing functions, I see `Record<string, unknown>` fits best for their purpose. Hopefully, turning on this flag didn't change the code a lot.