fix(linter): @typescript-eslint/unbound-method
For the static methods referenced, I put `this: void` annotation. It just explicitly removes a concern towards referencing invalid `this`. For `browser.webRequest.filterResponseData`, the type indicates that the method is static. However, I'm not sure and wrapping it with an arrow function looks like a safer approach since we just don't need to deal with it. Attaching `.bind(null)` is the thing I cannot explain.